Retry policy - grpc, sync
