+++ dapr/conf/__init__.py
@@ -23,7 +23,10 @@ class Settings:
         for setting in dir(global_settings):
             default_value = getattr(global_settings, setting)
             env_variable = os.environ.get(setting)
-            setattr(self, setting, env_variable or default_value)
+            if env_variable:
+                setattr(self, setting, type(default_value)(env_variable))
+            else:
+                setattr(self, setting, default_value)
 
     def __getattr__(self, name):
         if name not in dir(global_settings):
