+++ dapr/conf/global_settings.py
@@ -27,6 +27,9 @@ DAPR_GRPC_PORT = 50001
 DAPR_API_VERSION = 'v1.0'
 DAPR_HEALTH_TIMEOUT = 60  # seconds
 
+DAPR_API_MAX_RETRIES = 0
+DAPR_API_TIMEOUT_SECONDS = 60
+
 DAPR_API_METHOD_INVOCATION_PROTOCOL = 'http'
 
 DAPR_HTTP_TIMEOUT_SECONDS = 60
